+++ packages/liveblocks-core/src/types/Brand.ts
@@ -1,2 +1,0 @@
-declare const brand: unique symbol;
-export type Brand<T, TBrand extends string> = T & { [brand]: TBrand };